# Update `MiqReport::filter_with_report_results_by` to use join instead of include