@kbrock 2016-07-20
fixup:

```ruby
seed_array = [
  [10000, 7000, 6000],
  [400, 250, 60]*3,
  [20]*5, [10]*60
].flatten

miq_task_states   = ::MiqTask.validators.detect {|v| v.attributes.include? :state  }.options[:in]
miq_task_statuses = ::MiqTask.validators.detect {|v| v.attributes.include? :status }.options[:in]
miq_group         = ::MiqGroup.where(:description => "EvmGroup-user_self_service").first

seed_array.size.times do |report_result_count|
  name = "Test %04d" % (report_result_count+1)
  report = ::MiqReport.create! :name => name,
                               :title => name,
                               :db => 'test',
                               :miq_group_id => miq_group.id,
                               :rpt_group => 'test',
                               :rpt_type => 'Custom'

  report_result_count.times do |i|
    task = ::MiqTask.create!
    task.update_attributes! :state  => miq_task_states.sample.first,
                            :status => miq_task_statuses.sample.first

    ::MiqReportResult.create! :name => name,
                              :report => report,
                              :last_run_on => (i+1).day.ago,
                              :miq_task_id => task.id,
                              :miq_report_id => report.id,
                              :miq_group_id => miq_group.id
  end
end
```