@NickLaMuro 2016-07-20
@kbrock Yeah, sorry about that.  It was ripped from [miq_seeds](https://github.com/NickLaMuro/miq_seeds/blob/master/lib/miq_seeds/generators/miq_reports.rb), and just converted to a ruby script, so I didn't take the time to actually test it myself.  Hence why I abstracted this stuff away so I don't need to provide so much boilerplate for reviews.